+++ src/app/(main)/chat/settings/_layout/Desktop/Header.tsx
@@ -5,8 +5,9 @@ import { useRouter } from 'next/navigation';
 import { memo } from 'react';
 import { useTranslation } from 'react-i18next';
 
-import HeaderContent from '@/app/(main)/chat/settings/features/HeaderContent';
 import { pathString } from '@/utils/url';
+
+import HeaderContent from '../../features/HeaderContent';
 
 const Header = memo(() => {
   const { t } = useTranslation('setting');