+++ src/locales/default/auth.ts
@@ -1,0 +1,6 @@
+export default {
+  login: '登录',
+  loginOrSignup: '登录 / 注册',
+  signout: '退出登录',
+  signup: '注册',
+};