+++ src/types/fetch.ts
@@ -5,6 +5,7 @@ export const ChatErrorType = {
   // ******* 业务错误语义 ******* //
 
   InvalidAccessCode: 'InvalidAccessCode', // 密码无效
+  InvalidClerkUser: 'InvalidClerkUser', // is not Clerk User
   OpenAIBizError: 'OpenAIBizError', // OpenAI 返回的业务错误
   NoOpenAIAPIKey: 'NoOpenAIAPIKey',
   OllamaServiceUnavailable: 'OllamaServiceUnavailable', // 未启动/检测到 Ollama 服务